# papahana_flask_server_demo/papahana/models/container_summary.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from papahana.models.base_model_ import Model
from papahana import util


class ContainerSummary(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, name: str=None, semester: str=None, observation_blocks: List[str]=None, comment: str=None):  # noqa: E501
        """ContainerSummary - a model defined in Swagger

        :param id: The id of this ContainerSummary.  # noqa: E501
        :type id: str
        :param name: The name of this ContainerSummary.  # noqa: E501
        :type name: str
        :param semester: The semester of this ContainerSummary.  # noqa: E501
        :type semester: str
        :param observation_blocks: The observation_blocks of this ContainerSummary.  # noqa: E501
        :type observation_blocks: List[str]
        :param comment: The comment of this ContainerSummary.  # noqa: E501
        :type comment: str
        """
        self.swagger_types = {
            'id': str,
            'name': str,
            'semester': str,
            'observation_blocks': List[str],
            'comment': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'semester': 'semester',
            'observation_blocks': 'observation_blocks',
            'comment': 'comment'
        }
        self._id = id
        self._name = name
        self._semester = semester
        self._observation_blocks = observation_blocks
        self._comment = comment

    @classmethod
    def from_dict(cls, dikt) -> 'ContainerSummary':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ContainerSummary of this ContainerSummary.  # noqa: E501
        :rtype: ContainerSummary
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this ContainerSummary.


        :return: The id of this ContainerSummary.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this ContainerSummary.


        :param id: The id of this ContainerSummary.
        :type id: str
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this ContainerSummary.


        :return: The name of this ContainerSummary.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this ContainerSummary.


        :param name: The name of this ContainerSummary.
        :type name: str
        """

        self._name = name

    @property
    def semester(self) -> str:
        """Gets the semester of this ContainerSummary.


        :return: The semester of this ContainerSummary.
        :rtype: str
        """
        return self._semester

    @semester.setter
    def semester(self, semester: str):
        """Sets the semester of this ContainerSummary.


        :param semester: The semester of this ContainerSummary.
        :type semester: str
        """

        self._semester = semester

    @property
    def observation_blocks(self) -> List[str]:
        """Gets the observation_blocks of this ContainerSummary.


        :return: The observation_blocks of this ContainerSummary.
        :rtype: List[str]
        """
        return self._observation_blocks

    @observation_blocks.setter
    def observation_blocks(self, observation_blocks: List[str]):
        """Sets the observation_blocks of this ContainerSummary.


        :param observation_blocks: The observation_blocks of this ContainerSummary.
        :type observation_blocks: List[str]
        """

        self._observation_blocks = observation_blocks

    @property
    def comment(self) -> str:
        """Gets the comment of this ContainerSummary.


        :return: The comment of this ContainerSummary.
        :rtype: str
        """
        return self._comment

    @comment.setter
    def comment(self, comment: str):
        """Sets the comment of this ContainerSummary.


        :param comment: The comment of this ContainerSummary.
        :type comment: str
        """

        self._comment = comment