# papahana_flask_server_demo/papahana/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Papahana_demo
  description: WMKO API to the Observation Management System
  contact:
    email: observing-support@keck.hawaii.edu
  license:
    name: BSD 3-Clause
    url: https://opensource.org/licenses/BSD-3-Clause
  version: "0.1"
servers:
- url: http://localhost:8080/v0
...
paths:
...
  /containers:
    get:
      tags:
        - Containers
      description: Retrieves a specific container's information
      operationId: containers_get
      parameters:
        - $ref: '#/components/parameters/container_id'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Container'
        "400":
          description: Invalid status value
        "401":
          description: Unauthorized
        "404":
          description: Container not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
    put:
      tags:
        - Containers
      description: Overwrites a container
      operationId: containers_put
      parameters:
        - $ref: '#/components/parameters/container_id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Container'
        required: true
      responses:
        "201":
          description: OK
        "401":
          description: Unauthorized
        "404":
          description: Container id not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
    post:
      tags:
        - Containers
      description: Creates a container
      operationId: containers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Container'
        required: true
      responses:
        "201":
          description: OK
        "401":
          description: Unauthorized
        "404":
          description: Container id not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
    delete:
      tags:
        - Containers
      description: Delete container by id
      operationId: containers_delete
      parameters:
        - $ref: '#/components/parameters/container_id'
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Container id not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
  /containers/append:
    put:
      tags:
      - Containers
      description: Appends an observation block to a container by id
      operationId: containers_append_put
      parameters:
        - $ref: '#/components/parameters/container_id'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        "201":
          description: OK
        "400":
          description: Invalid status value
        "401":
          description: Unauthorized
        "404":
          description: Container not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
  /containers/executionTimes:
    get:
      tags:
      - Containers
      description: Calculate the total execution time of a container
      operationId: containers_execution_times_get
      parameters:
        - $ref: '#/components/parameters/container_id'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: number
                description: miliseconds
                x-content-type: application/json
        "400":
          description: Invalid status value
        "401":
          description: Unauthorized
        "404":
          description: Container not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
  /containers/items:
    get:
      tags:
      - Containers
      description: Retrieves the ordered list of observing blocks in a container
      operationId: containers_items_get
      parameters:
        - $ref: '#/components/parameters/container_id'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObservationBlock'
                x-content-type: application/json
        "400":
          description: Invalid status value
        "401":
          description: Unauthorized
        "404":
          description: Container not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
  /containers/items/summary:
    get:
      tags:
      - Containers
      description: Retrieves a summary of container information
      operationId: containers_items_summary_get
      parameters:
        - $ref: '#/components/parameters/container_id'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerSummary'
        "400":
          description: Invalid status value
        "401":
          description: Unauthorized
        "404":
          description: Container not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
  /containers/verify:
    get:
      tags:
      - Containers
      description: Request verification of the elements of a container. Sends container summary
        as a sucessfull response
      operationId: containers_verify_get
      parameters:
        - $ref: '#/components/parameters/container_id'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerSummary'
        "400":
          description: Invalid status value
        "401":
          description: Unauthorized
        "404":
          description: Container not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
  /containers/scheduleToO:
    post:
      tags:
      - Containers
      description: Submits a container for Target of Opportunity (ToO) (all the elements)
      operationId: containers_schedule_too_post
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Container'
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Invalid status value
        "401":
          description: Unauthorized
        "404":
          description: Container not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
  /containers/export:
    get:
      tags:
      - Containers
      description: Retrieves a specific container information in a file format (default
        .json)
      operationId: containers_export_get
      parameters:
        - $ref: '#/components/parameters/container_id'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Container'
        "400":
          description: Invalid status value
        "404":
          description: Container not found
      x-swagger-router-controller: Containers
      x-openapi-router-controller: papahana.controllers.containers_controller
...
  /obsBlocks/executionTime:
    get:
      tags:
        - Observation Block
      description: Calculates the execution time.
      operationId: ob_execution_time
      parameters:
        - $ref: '#/components/parameters/ob_id'
      responses:
        "200":
          description: On success returns an estimated execution time for an OB.
          content:
            application/json:
              schema:
                type: number
                x-content-type: application/json
        "400":
          description: Invalid status value
        "404":
          description: Observation Block not found
      x-swagger-router-controller: Observation
      x-openapi-router-controller: papahana.controllers.observation_block_controller
...
components:
  schemas:
...
    Container:
      type: object
      properties:
        id:
          type: object
        name:
          type: string
        semester:
          type: string
        observation_blocks:
          type: array
          items:
            type: string
        comment:
          type: string
      example:
        observation_blocks:
          - observation_blocks
          - observation_blocks
    ContainerSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        semester:
          type: string
        observation_blocks:
          type: array
          items:
            type: string
        comment:
          type: string
      example:
        container_id: container_id
    ObservationBlock:
      type: object
      properties:
        id:
          type: string
        signature:
          $ref: '#/components/schemas/Signature'
        target:
          type: array
          items:
            $ref: '#/components/schemas/Target'
        acquisition:
          $ref: '#/components/schemas/Acquisition'
        observations:
          type: array
          items:
            $ref: '#/components/schemas/Observation'
        associations:
          type: array
          items:
            type: string
        observation_type:
          type: array
          items:
            type: string
        priority:
          type: number
          example:
            99.99
        status:
          $ref: '#/components/schemas/Status'
        comment:
          type: string
          example:
            this is a nice ob block.
    Signature:
      type: object
      properties:
        name:
          type: string
        sem_id:
          type: string
        instrument:
          type: string
        pi_id:
          type: integer
        observer_ids:
          type: array
          items:
            type: integer
    Status:
      type: object
      properties:
        state:
          type: string
        executions:
          type: array
          items:
            type: string
    Acquisition:
      required:
      - acquisition
      - acquisition_method
      - ao_modes
      - associations
      - guider_selection
      - instrumental_setup
      - offset_stars
      - position_angles
      - slitmasks
      type: object
      properties:
        acquisition:
          type: string
        instrumental_setup:
          type: string
        acquisition_method:
          type: string
        guider_selection:
          type: string
        ao_modes:
          type: array
          items:
            type: string
        associations:
          type: array
          items:
            type: string
        offset_stars:
          type: array
          items:
            type: string
        slitmasks:
          type: array
          items:
            type: string
        position_angles:
          type: array
          items:
            type: string
      example:
        associations:
        - associations
        - associations
        offset_stars:
        - offset_stars
        - offset_stars
        position_angles:
        - position_angles
        - position_angles
        acquisition_method: acquisition_method
        guider_selection: guider_selection
        slitmasks:
        - slitmasks
        - slitmasks
        ao_modes:
        - ao_modes
        - ao_modes
        acquisition: acquisition
        instrumental_setup: instrumental_setup
    Observation:
      required:
      - associations
      - exposure_sequences
      - instrument
      type: object
      properties:
        instrument:
          type: string
        exposure_sequences:
          type: array
          items:
            type: string
        associations:
          type: array
          items:
            type: string
      example:
        associations:
        - associations
        - associations
        exposure_sequences:
        - exposure_sequences
        - exposure_sequences
        instrument: instrument
    Target:
      required:
      - ddec
      - dec
      - dec_offset
      - dra
      - epoch
      - equinox
      - frame
      - magnitude
      - name
      - obstime
      - pm_dec
      - pm_ra
      - ra
      - ra_offset
      type: object
      properties:
        name:
          type: string
        ra:
          type: string
        dec:
          type: string
        equinox:
          type: number
        frame:
          type: string
        ra_offset:
          type: number
        dec_offset:
          type: number
        pm_ra:
          type: number
        pm_dec:
          type: number
        epoch:
          type: number
        obstime:
          type: number
        magnitude:
          $ref: '#/components/schemas/Magnitude'
        wrap?:
          type: string
        dra:
          type: number
        ddec:
          type: number
        comment?:
          type: string
        template:
          type: boolean
      example:
        dec: dec
        epoch: 2.3021358869347655
        pm_dec: 5.637376656633329
        pm_ra: 5.962133916683182
        ra: ra
        ddec: 3.616076749251911
        ra_offset: 6.027456183070403
        dra: 9.301444243932576
        obstime: 7.061401241503109
        wrap?: wrap?
        comment?: comment?
        name: name
        magnitude:
          template: true
          magnitude: magnitude
          band: band
        dec_offset: 1.4658129805029452
        equinox: 0.8008281904610115
        frame: frame
    Magnitude:
      required:
      - band
      - magnitude
      - template
      type: object
      properties:
        band:
          type: string
        magnitude:
          type: string
        template:
          type: boolean
      example:
        template: true
        magnitude: magnitude
        band: band
...
  parameters:
...
    ob_id:
      name: ob_id
      in: query
      description: observation block id
      required: true
      style: form
      explode: true
      schema:
        type: string
      example: 6042ced0afc6ecd40fe5a3dd
...
    container_id:
      name: container_id
      in: query
      description: container identifier
      required: true
      explode: true
      schema:
        type: string
      example: 6042ced0afc6ecd40fe5a3dd
...